+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.8.0] -
+## Unreleased
+
+## [0.8.0] - 2018-02-05
 
 * Created `MultiSelectDelegate` and `MultiSelectCollectionViewController` based on [objc.io's Generic Table View Controllers](https://talk.objc.io/episodes/S01E26-generic-table-view-controllers-part-2)
 * Added `ItemsViewController` based on the same source as above
@@ -112,6 +114,7 @@ Older versions of this package were evolving at a really high pace.
 The package was not stable enough and maintaining a Changelog would have a real
 pain in the b***.
 
+[0.8.0]: https://github.com/FlaneurApp/FlaneurOpen/compare/0.7.0...0.8.0
 [0.7.0]: https://github.com/FlaneurApp/FlaneurOpen/compare/0.6.0...0.7.0
 [0.6.0]: https://github.com/FlaneurApp/FlaneurOpen/compare/0.5.0...0.6.0
 [0.5.0]: https://github.com/FlaneurApp/FlaneurOpen/compare/0.4.0...0.5.0